college_media/requirement.txt
django
pillow
channels_redis

